sessionAuth, sessionDestroy api and usersession class modification 17/05

+++ htdocs/src/classes/usersession.class.php
@@ -4,8 +4,7 @@
 class usersession
 {
     public $data;
-    public static function authenticate($user, $pass, $fingerprint)
-    {
+    public static function authenticate($user, $pass, $fingerprint){
         if(user::login($user, $pass)){ 
 
         $conn = database::getConnection();
@@ -41,29 +40,24 @@ class usersession
     }
 
 
-    public static function authorize ($token, $fingerprint)
-    {
+    public static function authorize ($token, $fingerprint){
         $session = new usersession($token);
         $host_ip = $_SERVER['REMOTE_ADDR'];
         $host_useragent = $_SERVER['HTTP_USER_AGENT'];
 
-        if($session->data['ip'] == $host_ip and $session->data['user_agent'] == $host_useragent and $session->data['fingerprint'] == $fingerprint)
-        {
+        if($session->data['ip'] == $host_ip and $session->data['user_agent'] == $host_useragent and $session->data['fingerprint'] == $fingerprint){
             return true;
-        }
-        else{
+        } else {
             return false;
         }
     }
 
-    public static function validate_session($token)
-    {
+    public static function validate_session($token){
         $session = new usersession($token);
         $host_ip = $_SERVER['REMOTE_ADDR'];
         $host_useragent = $_SERVER['HTTP_USER_AGENT'];
 
-        if($session->data['ip'] == $host_ip and $session->data['user_agent'] == $host_useragent)
-        {
+        if($session->data['ip'] == $host_ip and $session->data['user_agent'] == $host_useragent){
             session::$user = $session->getUser();
             return $session;
         }
@@ -72,9 +66,7 @@ class usersession
         }
     }
 
-public function __construct($token)
-{
-
+public function __construct($token){
     $conn = database::getConnection();
     $sql = "SELECT * FROM `sessions` WHERE `session_token`= '$token' LIMIT 1";
     $result = $conn->query($sql);
@@ -87,8 +79,7 @@ public function __construct($token)
     }
 }
 
-public function isValid()
-{
+public function isValid(){
     if (isset($this->data['login_time'])) {
         $login_time = DateTime::createFromFormat('Y-m-d H:i:s', $this->data['login_time']);
         if (3600 > time() - $login_time->getTimestamp()) {
@@ -101,8 +92,7 @@ public function isValid()
     }
 }
 
-public function getIP($token)
-{
+public function getIP($token){
     if(!$this->conn)
     {
         $this->conn = database::getConnection();
@@ -118,53 +108,59 @@ public function getIP($token)
     }
 }
 
-public function getUserAgent($token)
-{
-    if(!$this->conn)
-    {
-           $this->conn = database::getConnection();
-    }
- 
+public static function getUserAgent($token){
+    $conn = database::getConnection();
     $sql = "SELECT `user_agent` FROM `sessions` WHERE `session_token` = '$token'";
-    
-    $result = $this->conn->query($sql);
-    if ($result->num_rows) {
-        $row = $result->fetch_assoc();
-        return $this->ip = $row['user_agent'];
+    if ($conn->query($sql)->num_rows == 1) {
+        return $conn->query($sql)->fetch_assoc()['user_agent'];
     } else {
         return false;
     }
 }
 
-public function getUser()
-{
+public function getUser(){
     return new user($this->data['uid']);
 }
 
-public static function destroy($token)
-{ 
-    $conn = database::getConnection();
-    $sql = "DELETE FROM `sessions` WHERE `session_token` = '$token'";
-    try{
-        $result = $conn->query($sql);
-    }
-    catch(Exception $e)
-    {
-       return false;
+public static function destroy($token){ 
+    if(session::isAuthenticated()){
+        $conn = database::getConnection();
+        $sql = "DELETE FROM `sessions` WHERE `session_token` = '$token'";
+        try {
+            $result = $conn->query($sql);
+            session::destroy();
+            return true;
+        }
+        catch(Exception $e){
+        return false;
+        }
+    } else {
+        return false;
     }
     
-    if($result == true)
-    {
-        session::destroy();
-        return true;
-    }else{
-        session::destroy();
-        return false;
+}
+
+public static function validateSessionOwner($session_token){
+    $conn = database::getConnection();
+    $sql = "SELECT `ip` FROM `sessions` WHERE `session_token` = '$session_token'";
+    try{
+        if($conn->query($sql)->num_rows == 1){
+            $session_ip = $conn->query($sql)->fetch_assoc()['ip'];
+            if($_SERVER['REMOTE_ADDR'] == $session_ip and $_SERVER['HTTP_USER_AGENT'] == self::getUserAgent(session::get('session_token'))){
+                return true;
+            } else {
+                return false;
+            }
+        } else {
+            return false;
+        }
+    }
+    catch(Exception $e){
+    return false;
     }
 }
 
-public static function invalidate($token)
-{
+public static function invalidate($token){
     if(!$this->conn)
     {
            $this->conn = database::getConnection();
@@ -173,9 +169,8 @@ public static function invalidate($token)
     $result = $this->conn->query($sql);
     if($result)
     {
-        print "Session invalidated. Please Login again." . session_destroy(); ;
-
-}
+        print "Session invalidated. Please Login again." . session_destroy();
+    }
 }
